drivers/led: Extend userled_set_t from 8bits to 32bits

+++ boards/arm/lpc43xx/lpc4357-evb/src/lpc43_userleds.c
@@ -76,7 +76,7 @@
  *
  *  uint32_t board_userled_initialize(void);
  *  void board_userled(int led, bool ledon);
- *  void board_userled_all(uint8_t ledset);
+ *  void board_userled_all(uint32_t ledset);
  */
 
 /****************************************************************************
@@ -136,7 +136,7 @@ void board_userled(int led, bool ledon)
  * Name: board_userled_all
  ****************************************************************************/
 
-void board_userled_all(uint8_t ledset)
+void board_userled_all(uint32_t ledset)
 {
   lpc43_gpio_write(GPIO_LED, (ledset & BOARD_LED_BIT) == 0);
 }
